# Refactor article series items to inline orderable