--- a/templates/themes/ai_series/article_series_page.html
+++ b/templates/themes/ai_series/article_series_page.html
@@ -87,31 +87,29 @@
       <div class="row justify-content-center">
         <div class="col col-lg-10">
           <h2>In the Series</h2>
-          {% for block in self.series_items %}
-            {% define block.value.specific|page_type as page_type %}
-            {% if block.block_type == 'series_item' %}
-              <article>
-                {% image block.value.specific.image_banner max-450x200 as image_banner %}
-                <div class="background-image" style="background-image: url('{{image_banner.url}}');"></div>
-                <h3><a href="{{ block.value.url }}">{{ block.value.title }}</a></h3>
-                <div class="authors-container">
-                  {% if page_type == 'multimedia' %}
-                    <a class="video-button" href="{{ block.value.url  }}">
-                      <i class="fas fa-play-circle"></i>
-                    </a>
-                  {% endif %}
-                  <ul class="authors">
-                    {% for item in block.value.specific.authors.all %}
-                      {% if item.hide_link %}
-                        <li>{{ item.author.title }}</li>
-                      {% else %}
-                        <li><a href="{% pageurl item.author %}">{{ item.author.title }}</a></li>
-                      {% endif %}
-                    {% endfor %}
-                  </ul>
-                </div>
-              </article>
-            {% endif %}
+          {% for series_item in self.article_series_items %}
+            {% define series_item.content_page.specific|page_type as page_type %}
+            <article>
+              {% image series_item.content_page.specific.image_banner max-450x200 as image_banner %}
+              <div class="background-image" style="background-image: url('{{image_banner.url}}');"></div>
+              <h3><a href="{% pageurl series_item.content_page %}">{{ series_item.content_page.title }}</a></h3>
+              <div class="authors-container">
+                {% if page_type == 'multimedia' %}
+                  <a class="video-button" href="{% pageurl series_item.content_page %}">
+                    <i class="fas fa-play-circle"></i>
+                  </a>
+                {% endif %}
+                <ul class="authors">
+                  {% for item in series_item.content_page.authors.all %}
+                    {% if item.hide_link %}
+                      <li>{{ item.author.title }}</li>
+                    {% else %}
+                      <li><a href="{% pageurl item.author %}">{{ item.author.title }}</a></li>
+                    {% endif %}
+                  {% endfor %}
+                </ul>
+              </div>
+            </article>
           {% endfor %}
         </div>
       </div>