--- a/templates/themes/data_series/article_series_page.html
+++ b/templates/themes/data_series/article_series_page.html
@@ -72,23 +72,22 @@
       <div class="container">
         <div class="row">
           <div class="col">
-            {% for block in self.series_items %}
-              {% if block.block_type == 'series_item' %}
-                <article>
-                  <h4 class="article-title"><a href="{{ block.value.url }}">{{ block.value.title}}</a></h4>
-                  <p class="authors">
-                    {% for item in block.value.specific.authors.all %}
-                      {% if item.hide_link %}
-                        <span>{{ item.author.title }}</span>
-                      {% else %}
-                        <span><a href="{% pageurl item.author %}">{{ item.author.title }}</a></span>
-                      {% endif %}
-                    {% endfor %}
-                  </p>
-                </article>
-              {% elif block.block_type == 'category_title' %}
-                <h3 class="section-title">{{ block.value }}</h3>
+            {% for series_item in self.article_series_items %}
+              {% if series_item.category_title %}
+                <h3 class="section-title">{{ series_item.category_title }}</h3>
               {% endif %}
+              <article>
+                <h4 class="article-title"><a href="{% pageurl series_item.content_page %}">{{ series_item.content_page.title}}</a></h4>
+                <p class="authors">
+                  {% for item in series_item.content_page.authors.all %}
+                    {% if item.hide_link %}
+                      <span>{{ item.author.title }}</span>
+                    {% else %}
+                      <span><a href="{% pageurl item.author %}">{{ item.author.title }}</a></span>
+                    {% endif %}
+                  {% endfor %}
+                </p>
+              </article>
             {% endfor %}
           </div>
         </div>