Performance improvement:
Page | Before - Time | Before - Queries | After - Time | After - Queries
--- | --- | --- | --- | ---
Opinions | 52200.82ms | 817 queries in 1204.29ms | 12786.95ms | 203 queries in 242.81ms
Modern Conflict and Artificial Intelligence | 11186.26ms | 119 queries in 91.03ms | 3809.71ms | 55 queries in 854.67ms
Security, Intelligence and the Global Health Crisis | 12228.68ms | 186 queries in 149.40ms | 3927.72ms | 73 queries in 226.42ms
Global Cooperation after COVID-19 | 17145.88ms | 237 queries in 241.20ms | 4438.65ms | 89 queries in 252.34ms
Modernizing the World Trade Organization | 12511.37ms | 94 queries in 119.07ms | 3408.73ms | 42 queries in 72.36ms
Models for Platform Governance | 25900.19ms | 139 queries in 133.49ms | 3273.32ms | 69 queries in 151.76ms